backend/requirements.txt
python-socketio[asyncio_redis]
uvicorn
ollama